Randall || hamburger menu
Added a bunch to the beginning of the html file mostly. Added a wrapper
to the entire body. I needed to add a sticky header for the menu to sit
on. Savannah, you can use this sticky header and make it pretty, or you
can make your own and I will put my menu on that. As far as CSS goes, I
added some normalize stuff and other things that only affect the  menu,
header, and wrapper. Let me know if you think this menu should do
anything else: smaller, bigger, bright orange etc.. Thanks!